#88 documentation much improved.
